refactor: 중복된 PageResponse를 제거하라
