+++ src/main/java/prgrms/project/stuti/domain/feed/repository/postcomment/PostCommentCustomRepository.java
@@ -1,7 +1,7 @@
 package prgrms.project.stuti.domain.feed.repository.postcomment;
 
 import prgrms.project.stuti.domain.feed.service.dto.CommentParentContents;
-import prgrms.project.stuti.global.page.offset.PageResponse;
+import prgrms.project.stuti.global.page.PageResponse;
 
 public interface PostCommentCustomRepository {
