+++ src/main/java/prgrms/project/stuti/domain/feed/service/PostCommentConverter.java
@@ -10,7 +10,7 @@ import prgrms.project.stuti.domain.feed.service.response.PostCommentContentsResp
 import prgrms.project.stuti.domain.feed.service.dto.PostCommentGetDto;
 import prgrms.project.stuti.domain.feed.service.response.PostCommentResponse;
 import prgrms.project.stuti.domain.member.model.Member;
-import prgrms.project.stuti.global.page.offset.PageResponse;
+import prgrms.project.stuti.global.page.PageResponse;
 
 public class PostCommentConverter {
