--- src/main/java/prgrms/project/stuti/global/page/offset/PageResponse.java
@@ -1,10 +1,0 @@
-package prgrms.project.stuti.global.page.offset;
-
-import java.util.List;
-
-public record PageResponse <T>(
-	List<T> contents,
-	boolean hasNext,
-	Long totalElements
-) {
-}
