+++ src/test/java/prgrms/project/stuti/domain/feed/controller/PostCommentControllerTest.java
@@ -23,7 +23,7 @@ import prgrms.project.stuti.domain.feed.service.PostCommentService;
 import prgrms.project.stuti.domain.feed.service.dto.CommentParentContents;
 import prgrms.project.stuti.domain.feed.service.response.PostCommentContentsResponse;
 import prgrms.project.stuti.domain.feed.service.response.PostCommentResponse;
-import prgrms.project.stuti.global.page.offset.PageResponse;
+import prgrms.project.stuti.global.page.PageResponse;
 
 @WebMvcTest(PostCommentController.class)
 class PostCommentControllerTest extends TestConfig {
